move non-algorithm file into detail directory
