+++ core/include/algorithms/seeding/detail/seeding_config.hpp
